# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API sepcifications for prepper",
        "title": "prepper API",
        "contact": {},
        "version": "0.1"
    },
    "paths": {
        "/ai/analyse": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai"
                ],
                "summary": "analyse an answer to a question",
                "parameters": [
                    {
                        "description": "Question and Answer to be analysed by AI",
                        "name": "QAPair",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AnalysisRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.Analysis"
                        }
                    }
                }
            }
        },
        "/ai/coverletter": {
            "post": {
                "description": "Inputs include given name, education, position and company you are applying for, reasons to apply and previous experience",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai"
                ],
                "summary": "creates personalised cover letter",
                "parameters": [
                    {
                        "description": "cover letter request",
                        "name": "CoverLetterDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CoverLetterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.CoverLetterResponse"
                        }
                    }
                }
            }
        },
        "/ai/text2voice": {
            "post": {
                "description": "Produces an audio file reading out the given text",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ai"
                ],
                "summary": "converts text to voice",
                "responses": {}
            }
        },
        "/ai/voice2text": {
            "post": {
                "description": "The API endpoint expects an audio file in audio/webm format to be provided in the request body as a blob.",
                "consumes": [
                    "application/octet-stream"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai"
                ],
                "summary": "convert voice to text",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.Voice2TextResponse"
                        }
                    }
                }
            }
        },
        "/application": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "Delete a user's application",
                "parameters": [
                    {
                        "description": "Application",
                        "name": "ApplicationDeleteRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ApplicationDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/application/:applicationId/questions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "Get an user's application question types",
                "parameters": [
                    {
                        "type": "string",
                        "description": "applicationId",
                        "name": "applicationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApplicationQuestionResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "test"
                    }
                }
            }
        },
        "/application/:applicationId/questions/generate": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "Use AI to generate questions based on questions tags in a specified application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "applicationId",
                        "name": "applicationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.GeneratedQuestion"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "test"
                    }
                }
            }
        },
        "/application/create": {
            "post": {
                "description": "an application has some properties",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "Create an application for the user",
                "parameters": [
                    {
                        "description": "Application",
                        "name": "ApplicationCreateBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ApplicationCreateBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ApplicationCreateResponse"
                        }
                    },
                    "400": {
                        "description": "failed to create app"
                    }
                }
            }
        },
        "/application/me": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "Get user's applications",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.InnerApplication"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "ping"
                ],
                "summary": "Ping the server",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/healthcheck": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Check if a user is signed in",
                "responses": {
                    "200": {
                        "description": "if signed in `{\"success\": true}`"
                    },
                    "401": {
                        "description": "if not signed in"
                    }
                }
            }
        },
        "/user/signin": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Sign a user",
                "parameters": [
                    {
                        "description": "Password must be atleast 6 characters.",
                        "name": "SignInBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserCredentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserSigninResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid Credentials"
                    }
                }
            }
        },
        "/user/signout": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Sign a user out of dancord",
                "responses": {
                    "200": {
                        "description": "on successful signout"
                    }
                }
            }
        },
        "/user/signup": {
            "post": {
                "description": "Insert description here",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Sign a user up to dancord",
                "parameters": [
                    {
                        "description": "Password must be atleast 6 characters.",
                        "name": "SignUpBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserCredentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserSigninResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ApplicationCreateBody": {
            "type": "object",
            "required": [
                "jobDescription",
                "name"
            ],
            "properties": {
                "jobDescription": {
                    "type": "string",
                    "example": "Looking for an engineer to join our team."
                },
                "name": {
                    "type": "string",
                    "example": "SafetyCulture"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "ApplicationCreateResponse": {
            "type": "object",
            "required": [
                "jobDescription",
                "name"
            ],
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2021-07-01T00:00:00.000Z"
                },
                "id": {
                    "type": "string",
                    "example": "1337"
                },
                "jobDescription": {
                    "type": "string",
                    "example": "Looking for an engineer to join our team."
                },
                "name": {
                    "type": "string",
                    "example": "SafetyCulture"
                }
            }
        },
        "UserCredentials": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "daniel321"
                },
                "username": {
                    "type": "string",
                    "example": "daniel"
                }
            }
        },
        "UserSigninResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "1337"
                }
            }
        },
        "db.InnerApplication": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jobDescription": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ownerId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "handlers.AnalysisRequest": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                }
            }
        },
        "handlers.ApplicationDeleteRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "handlers.ApplicationQuestionResponse": {
            "type": "object",
            "properties": {
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.QuestionType"
                    }
                }
            }
        },
        "handlers.CoverLetterRequest": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "education": {
                    "type": "string"
                },
                "experience": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "reasons": {
                    "type": "string"
                }
            }
        },
        "handlers.CoverLetterResponse": {
            "type": "object",
            "properties": {
                "coverLetter": {
                    "type": "string"
                }
            }
        },
        "handlers.GeneratedQuestion": {
            "type": "object",
            "properties": {
                "audioLink": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "questionPrompt": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "handlers.QuestionType": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "service.Analysis": {
            "type": "object",
            "properties": {
                "bad": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "good": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "service.Voice2TextResponse": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ]
}